learn_about_people/requirements.txt
speech @ git+https://github.com/viam-labs/speech.git@main
viam-sdk
pillow